Move Bugzilla models under Bugzilla namespace
