+++ tests/unit/bugzilla/test_client.py
@@ -3,7 +3,7 @@ import responses
 from responses import matchers
 
 from jbi.bugzilla.client import BugzillaClient, BugzillaClientError
-from jbi.models import BugzillaWebhookRequest
+from jbi.bugzilla.models import BugzillaWebhookRequest
 
 
 @pytest.fixture
