+++ tests/unit/test_app.py
@@ -5,8 +5,8 @@ import pytest
 from fastapi.testclient import TestClient
 
 from jbi.app import app, traces_sampler
+from jbi.bugzilla.models import BugzillaWebhookRequest
 from jbi.environment import get_settings
-from jbi.models import BugzillaWebhookRequest
 
 
 def test_request_summary_is_logged(caplog, anon_client):
