+++ tests/unit/test_runner.py
@@ -6,9 +6,10 @@ import requests
 import responses
 
 from jbi import Operation
+from jbi.bugzilla.models import BugzillaWebhookRequest
 from jbi.environment import get_settings
 from jbi.errors import IgnoreInvalidRequestError
-from jbi.models import ActionContext, BugzillaWebhookRequest
+from jbi.models import ActionContext
 from jbi.runner import Executor, execute_action
